manual test removed from models/cart.js into tests/cart.manual.js
